+++ experiment/src/org/labkey/experiment/lineage/LineageDisplayColumn.java
@@ -267,16 +267,16 @@ public class LineageDisplayColumn extends DataColumn implements IMultiValuedDisp
         static final String OBJECTID_PARAMETER = "_$_OBJECTID_$_";
         final UserSchema schema;
         final SQLFragment sqlf;
-        final List<QueryService.ParameterDecl> parameters = Collections.singletonList(new QueryService.ParameterDeclaration(OBJECTID_PARAMETER, JdbcType.INTEGER));
+        final List<QueryService.ParameterDecl> parameters = Collections.singletonList(new QueryService.ParameterDeclaration(OBJECTID_PARAMETER, JdbcType.BIGINT));
 
         SeedTable(UserSchema schema, @Nullable ContainerFilter cf)
         {
             super(schema.getDbSchema(), "seed");
             SqlDialect d = schema.getDbSchema().getScope().getSqlDialect();
             this.schema = schema;
-            this.sqlf = new SQLFragment("SELECT CAST( ? AS " + d.getSqlCastTypeName(JdbcType.INTEGER)+ ") AS objectid");
+            this.sqlf = new SQLFragment("SELECT CAST( ? AS " + d.getSqlCastTypeName(JdbcType.BIGINT)+ ") AS objectid");
             this.sqlf.addAll(parameters);
-            var objectidCol = new BaseColumnInfo("objectid", this, JdbcType.INTEGER);
+            var objectidCol = new BaseColumnInfo("objectid", this, JdbcType.BIGINT);
             addColumn(objectidCol);
             var inputs = new AliasedColumn(this, "Inputs", objectidCol);
             inputs.setFk(LineageForeignKey.createWithMultiValuedColumn(schema, sqlf, true, cf));
